requirements.txt
pillow~=10.4.0
torch~=2.10.0
matplotlib~=3.10.8
torchvision~=0.25.0
scikit-learn~=1.7.2
fastapi~=0.129.0
uvicorn~=0.41.0